refactor : RacingCarGame을 메인으로 변경 <description> 기존의 index.js보다 RacingCarGame안에서 게임을 시작하는 것이 가독성이 더 좋아보여서 변경
